+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * (feature) Extract parsing of `_editable` data into `PreviewDataParser`.
 * (improvement) Expose `_editable` data via `StoryMetaData::getPreviewData()`.
+* (feature) Add `ComponentPreviewData` helper to easily render component meta and preview data.
 
 
 3.2.2